Tweak for ffmpeg thumbnail, better frames and more clips (#5)
## About 
Using information found in docs, https://ffmpeg.org/ffmpeg-all.html#thumbnail, able to modify how many frames FFmpeg uses to find a frame for a thumbnail. Set steps to 250, above default 100, and was able to get better results.

Also changed how many clips we are generating, we are doing about 30 clips for a video, instead of 25. Seems to be getting better results with GPT and just looking at the frame strips.

## Notes
* https://ffmpeg.org/ffmpeg-all.html#thumbnail
* https://superuser.com/questions/1336285/use-ffmpeg-for-thumbnail-selections