Creating Chat API Service (#9)
== Overview ==
Creating Chat API Service to interact with embed database and return chat results to a user base on that user prompts. API is design as a REST service, outline how each call is in spec.md

== Note ==
Haven't added queue system using BullMQ and Redis. Need to add that in order to have better scale at launch.